Build fix.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@67555 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Audio/DevHDA.cpp
@@ -6337,7 +6337,7 @@ static int hdaLoadExecLegacy(PHDASTATE pThis, PSSMHANDLE pSSM, uint32_t uVersion
             if (RT_FAILURE(rc))
                 break;
 
-            cStreams = RT_CLAMP(cStreams, 0, HDA_MAX_STREAMS); /* Sanity. */
+            cStreams = RT_CLAMP((int32_t)cStreams, 0, HDA_MAX_STREAMS); /* Sanity. */
 
             /* Load stream states. */
             for (uint32_t i = 0; i < cStreams; i++)
@@ -6536,7 +6536,7 @@ static DECLCALLBACK(int) hdaLoadExec(PPDMDEVINS pDevIns, PSSMHANDLE pSSM, uint32
     rc = SSMR3GetU32(pSSM, &cStreams);
     AssertRC(rc);
 
-    cStreams = RT_CLAMP(cStreams, 0, HDA_MAX_STREAMS); /* Sanity. */
+    cStreams = RT_CLAMP((int32_t)cStreams, 0, HDA_MAX_STREAMS); /* Sanity. */
 
     Log2Func(("cStreams=%RU32\n", cStreams));
